refactor: extract chat generation state setup
