+++ Roadmap.md
@@ -882,6 +882,7 @@ Done and tested:
   - `useGenerationStreamUpdate.js` now owns stream fan-out and throttled background DB persistence during active generation
   - `useGenerationPromptReady.js` now owns prompt metadata assignment and prompt-ready persistence for triggered lorebooks/memories/context refs
   - `useGenerationPreparation.js` now owns authors-note assembly, placeholder message creation, and request history shaping for chat generation
+  - `useGenerationStateSetup.js` now owns initial guidance patching plus per-generation UI update/timer state wiring
 
 Tested status:
 - [done] `npm test -- --run`
